pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["denis <yudaev1@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.9"
python-dotenv = "^0.19.2"
gunicorn = "^20.1.0"
django-bootstrap4 = "^21.1"
whitenoise = "^5.3.0"
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.9.2"
django-filter = "^21.1"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
coverage = "^6.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
